*) core/mod_http1: remove some http1 filters from mod_core.h again,      making them non-public. Do some renames for clarity of purpose      what a filter does.      Use post_read_request hook to install http1 specific filters      when needed, removing knowledge from generic server/protocol.c      handling.
