+++ modules/http1/mod_http1.h
@@ -22,3 +22,40 @@
  */
 int http1_set_keepalive(request_rec *r, ap_bucket_headers *resp);
 
+/**
+ * Context for writing out HTTP/1 protocol fields, such as headers.
+ * Passed in iterators, for example.
+ */
+typedef struct http1_out_ctx_t {
+    apr_pool_t *pool;
+    apr_bucket_brigade *bb;
+} http1_out_ctx_t;
+
+/**
+ * Send a single HTTP header field to the client.  Note that this function
+ * is used in calls to apr_table_do(), so don't change its interface.
+ * It returns true unless there was a write error of some kind.
+ */
+int http1_write_header_field(http1_out_ctx_t *out,
+                             const char *fieldname, const char *fieldval);
+
+/**
+ * Write a response `resp` for request `r` in HTTP/1.x format to brigade `bb`.
+ */
+void http1_write_response(request_rec *r,
+                          ap_bucket_headers *resp,
+                          apr_bucket_brigade *bb);
+
+/* This is the HTTP1_BODY_IN filter for HTTP/1.x requests
+ * and responses from proxied servers (mod_proxy).
+ * It handles chunked and content-length bodies. This can only
+ * be inserted/used after a request has been read on a
+ * HTTP/1.x connection.
+ */
+apr_status_t http1_body_in_filter(ap_filter_t *f,
+                                  apr_bucket_brigade *b,
+                                  ap_input_mode_t mode,
+                                  apr_read_type_e block,
+                                  apr_off_t readbytes);
+
+
